moves merge code to new subdir
