+++ merge/merge_sheets.py
@@ -5,7 +5,7 @@ import time
 
 from SheetConfig import SheetConfig
 from exceptions import SheetConfigException, SheetManagerException
-from SheetManager import SheetManager
+from merge.SheetManager import SheetManager
 
 if __name__ == '__main__':
     parser = ArgumentParser()
